Splash + aRGB colors

+++ Chili Framework 2016/Engine/Graphics.h
@@ -75,11 +75,11 @@ public:
 		DrawRect(x, y,  x2,  y2, { unsigned char(r),unsigned char(g),unsigned char(b) });
 	}
 	void DrawRect(int x1, int y1, int x2, int y2, Color c);
-	void DrawBitmap(const uint8_t bitmap[], int x, int y, int width, int height);
+	void DrawBitmap(const unsigned int bitmap[], int x, int y, int width, int height);
 
-	void DrawCharacter(const uint8_t bitmap[], int offset, int x, int y, int width, int height);
+	void DrawCharacter(const unsigned int bitmap[], int offset, int x, int y, int width, int height);
 
-	void WriteLine(char* text, const uint8_t bitmap[], int x, int y, int charWidth, int charHeight);
+	void WriteLine(char* text, const unsigned int bitmap[], int x, int y, int charWidth, int charHeight);
 
 	~Graphics();
 private:
